Neovim: Replace the vim-markdown plugin
The vim-markdown plugin did weird things that caused list formatting to
no longer work. This other version is a stripped down fork that only
maintains the synax highlighting. This is also the primary feature
that caused me to select this plugin (support for LaTeX in `$$ $$`
blocks, so the lack of the other features is no problem.