fontconfig: Use Fira Code instead of Fira Mono
Fira Code is a monospaced font with programming ligatures. It is based
on Fira Mono and therefore very similar.

Ligatures are an interesting feature, as they visually change certain
character combinations to symbols, without changing the underlying
characters themselves. Setting it as the default font through fontconfig
will enable all application to use this font. Not all application
support ligatures however, but it seems that they will simply ignore
this, making it safe to set globally.