[fish] Restructure Ansible role
As an experiment, restructure the structure and task file of the Ansible
role for Fish. This new structure assumes that all files stored in the
'files/templates' directory are a template which has to be copied (even
if they to not contain variables). This makes it easier to add new
files, and add variables where desired.

A second modification is that unmanaged files will be removed. This
will make it easier to modify the configuration structure in the future.
If certain files are renamed, they will automatically be cleaned up (and
not linger and cause strange problems).