Move creating of config files to _common role
Creating the configuration files is largely the same for most roles;
directories are created and config files rendered from the templates.
Since this is the same for most roles, it can be moved to the _common
role. This reduces the boilerplate for many roles to simply including
the _common role.