[neovim] Configure Neovim to use packer.nvim
The plugin manager packer.nvim is a relatively new plugin manager for
Neovim. It is written in Lua and requires the recently released Neovim
version 0.5. Because it uses Lua instead of Vimscript, configuration can
be more expression and possible faster. This plugin manger can manage
plugins written in Lua as well as Vimscript.