[neovim] Don't manage plugins with Git submodules
Instead of using Git submodules, use a plugin manager for all plugins.
Using Git submodules was an experiment in using the native package
manager. However, Git submodules are not as flexible as declaring
a single line in a configuration file. Updates are more cumbersome, and
deleting a submodule is not very straightforward.