[neovim] Use recent version of treesitter plugin
The treesitter plugin `nvim-treesitter` was still pinned on the
`0.5-compat` branch, but that is no longer necessary since Neovim
version 0.7 is currently in use. This restriction can therefore be
lifted, and the most recent version of this plugin can be used.

There has been a breaking change, which deprecated the "maintained"
value for `ensure_installed`. Instead, "all" can be used to prevent
manually specifying the desired parsers.