[neovim] Add compatibility mappings <C-BS> <C-/>
The defined mappings <C-BS> and <C-/> no longer worked in recent version
of Neovim, because of a change in how keys are  registered. Ironically,
keys are now registered correctly, and a new mapping has been added. The
"incorrect" mappings must still be kept to remain compatible with tools
such tmux.