[meta] Add schema to with_filetree in Ansible task
The lookup plugin `with_filetree` has been moved to the
`community.general` collection. Also, Ansible Lint expects that the
value specified for `with_community.general.filetree` must be
variable of template. Simply passing setting it to `data/` gives the
following error:

```
deprecated-bare-vars: Possible bare variable 'data/' used in a 'with_community.general.filetree' loop. You should use the full variable syntax ('{{ data/ }}') or convert it to a list if that is not really a variable.
```

Converting it to a list also doesn't work, this still gives the same
error. This 'problem' is known, but it is not considered a bug:
https://github.com/ansible/ansible-lint/issues/1621

The suggested workaround is therefore used: putting the value as
a string inside a jinja2 template.