[fish] Do not modify $PATH directly
This approach of adding a value directly to `$PATH`, has the result that
nested `fish` instances have the same element added multiple times.

Instead, use `fish_add_path`. This function checks if the value is
already present, and will not add it again if this is the case. Use the
`--path` option, so `$PATH` is modified instead of adding it to
`fish_user_paths`. The latter makes it difficult to prepend an element
to `$PATH`.