Add colors/theme definition in Scheme
Create a Scheme file that returns the same data as that defined in
`theme.yml`. Instead of variable expansion with Jinja2 and Ansible, use
Scheme data types and variables.

This file can be used to generate the exact same data, also in order to
use this in Guix Home configuration.