Replace theme.yml with file generated theme.json
The theme definition in the Scheme file `theme.scm` is functionally
identical to the Jinja2/Ansible variable file `theme.yml`. It makes
sense to generate one of them based on the other.

Since YAML is a subset of JSON, any valid JSON is also valid YAML.
Generating a JSON file from `theme.yml` is straightforward, and easier
than doing it the other way around.