[fzf] Create separate Guix Home managed fzf config
Instead of placing the config of fzf and the installation of its
dependencies in the Ansible role for Fish, create a separate config that
is managed by Guix Home.

Add two user packages: fzf and fd. These will be installed in the Home
profile, and the binaries from these packages will now be used by
default.

Define two services: one that sets the environment variables used to
configure fzf, and another that configures Fish to set up the fzf key
bindings. Using two services like this would not have been possible
without the refactor of how services are loaded.

Now that fzf is managed by Guix Home, remove the relevant parts from the
Ansible role.