Add script to compute Guix templates checksum
This will read all the template files used by Guix services, compute the
checksum, and store this in `templates.checksum`.

This file can be used to determine if there has been a change in the
templates, including a removed file. This file can be used as a Makefile
dependency, as it will only be updated if the checksum has changed.

There are two Makefile targets: one for the actual file, which can be
used a dependency for other targets, and one `.PHONY` target which can
be used to always generate the checksum.