[waybar] Add initial Nix module with empty config
Set no configuration, except for the position of the main bar. This
means the bar will have no modules. Also set the styling to an empty
string, to revert to the default styles. The result is a 1-2 pixel white
line at the top of the screen.

Enable systemd integration, so `waybar` will be started as a systemd
service, as soon as the graphical user environment is initialized.