[neovim] Remove Pandoc plugins and settings
The Pandoc plugins are no longer needed, because the standard markdown
features are sufficient now. Syntax highlight has improved because of
Treesitter; math blocks no longer mess up the formatting of the rest of
the document and YAML frontmatter detected properly. One of the
downsides is that LaTeX syntax is no longer highlighted.

Another benefit is that all other Markdown settings are automatically
used instead of the Pandoc filetype. For example, `#` is no longer seen
as a 'comment', which would insert another `#` on a newline.

Certain features of the Pandoc plugin where disabled, others were not
being used. If they are missed, adding it again is always possible.