[gnome] Replace quake-terminal with ddterm
The main benefit of quake-terminal was that is offered the ability to
use your preferred terminal emulator. It did however have some
downsides. It would sometimes 'forget' the instance it was managing, and
then it would use another already open terminal instance instead. The
original instance would then be unreachable. It would sometimes also
only open on the same workspace, so opening the dropdown would move you
to this workspace.

ddterm offers its own terminal emulator, but otherwise it works fine so
far.